+++ src/scnnoise/gillespieSSA.hpp
@@ -82,7 +82,7 @@ namespace ScnnoiseInterface {
         void write_statistics_to_file (std::string statistics_file);
 
         void save_molecule_count_at_random_times (double time_prev, double time_next,
-            double &which_random_time_saved);
+            unsigned int &which_random_time_saved);
     };
 }
